Updating to work with Go 1.19.3
