+++ .github/workflows/main.yml
@@ -7,7 +7,7 @@ jobs:
     timeout-minutes: 30
     strategy:
       matrix:
-        go-version: [1.16.15]
+        go-version: [1.19.3]
         os: [ubuntu-latest]
     steps:
       - name: Checkout repository
