+++ client/src/worker/wasm_exec.d.ts
@@ -1,4 +1,5 @@
 // NB: tried @types/golang-wasm
+
 declare class Go {
   constructor();
   importObject: Record<string, Record<string, any>>;
